+++ backend/src/main/java/org/cathori/backend/notice/api/NoticeController.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.cathori.backend.notice.api.dto.NoticeDetailResponse;
 import org.cathori.backend.notice.api.dto.NoticeFeedResponse;
 import org.cathori.backend.notice.api.dto.NoticeSearchResponse;
-import org.cathori.backend.notice.application.NoticeFeedService;
+import org.cathori.backend.notice.application.query.NoticeFeedService;
 import org.cathori.backend.security.CustomUserDetails;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
