+++ backend/src/main/java/org/cathori/backend/notice/application/crawling/NoticeService.java
@@ -1,7 +1,8 @@
-package org.cathori.backend.notice.application;
+package org.cathori.backend.notice.application.crawling;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.cathori.backend.notice.application.AiPort;
 import org.cathori.backend.notice.infra.NoticeSummaryUpdater;
 import org.cathori.backend.notice.infra.ai.AiSummaryResult;
 import org.cathori.backend.notice.model.Notice;
