+++ backend/src/main/java/org/cathori/backend/notice/application/query/BookmarkedNoticeQuery.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.notice.application;
+package org.cathori.backend.notice.application.query;
 
 /**
  * 로그인 사용자의 북마크 공지 목록 조회 조건.
