+++ backend/src/main/java/org/cathori/backend/notice/application/query/NoticeFeedQuery.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.notice.application;
+package org.cathori.backend.notice.application.query;
 
 import java.util.List;
