+++ backend/src/main/java/org/cathori/backend/notice/infra/RetryScheduler.java
@@ -1,7 +1,7 @@
 package org.cathori.backend.notice.infra;
 
 import lombok.RequiredArgsConstructor;
-import org.cathori.backend.notice.application.NoticeService;
+import org.cathori.backend.notice.application.crawling.NoticeService;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
