+++ backend/src/main/java/org/cathori/backend/notice/model/Notice.java
@@ -5,7 +5,7 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.cathori.backend.notice.application.CrawledNotice;
+import org.cathori.backend.notice.application.crawling.CrawledNotice;
 import org.cathori.backend.notice.infra.ai.AiSummaryResult;
 import tools.jackson.core.JacksonException;
 import tools.jackson.databind.ObjectMapper;
