+++ backend/src/test/java/org/cathori/backend/notice/api/NoticeFeedIntegrationTest.java
@@ -3,7 +3,7 @@ package org.cathori.backend.notice.api;
 import org.cathori.backend.IntegrationTestBase;
 import org.cathori.backend.bookmark.domain.Bookmark;
 import org.cathori.backend.bookmark.infra.BookmarkJpaRepository;
-import org.cathori.backend.notice.application.CrawledNotice;
+import org.cathori.backend.notice.application.crawling.CrawledNotice;
 import org.cathori.backend.notice.infra.ai.AiSummaryResult;
 import org.cathori.backend.notice.model.Notice;
 import org.cathori.backend.notice.model.NoticeRepository;
