+++ backend/src/test/java/org/cathori/backend/notice/api/NoticeSearchIntegrationTest.java
@@ -1,7 +1,7 @@
 package org.cathori.backend.notice.api;
 
 import org.cathori.backend.IntegrationTestBase;
-import org.cathori.backend.notice.application.CrawledNotice;
+import org.cathori.backend.notice.application.crawling.CrawledNotice;
 import org.cathori.backend.notice.infra.ai.AiSummaryResult;
 import org.cathori.backend.notice.model.Notice;
 import org.cathori.backend.notice.model.NoticeRepository;
