+++ backend/src/test/java/org/cathori/backend/user/api/UserIntegrationTest.java
@@ -5,7 +5,7 @@ import org.cathori.backend.notification.domain.AlertHistory;
 import org.cathori.backend.notification.infra.AlertHistoryJpaRepository;
 import org.cathori.backend.bookmark.domain.Bookmark;
 import org.cathori.backend.bookmark.infra.BookmarkJpaRepository;
-import org.cathori.backend.notice.application.CrawledNotice;
+import org.cathori.backend.notice.application.crawling.CrawledNotice;
 import org.cathori.backend.notice.model.Notice;
 import org.cathori.backend.notice.model.NoticeRepository;
 import org.cathori.backend.security.JwtUtil;
